chore: Apply PHPStan Rules Lvl 6 WIP

+++ src/Controller/CallbackSettingsController.php
@@ -9,8 +9,8 @@
  * Full copyright and license information is available in
  * LICENSE.md which is distributed with this source code.
  *
- *  @copyright  Copyright (c) elements.at New Media Solutions GmbH (https://www.elements.at)
- *  @license    http://www.pimcore.org/license     GPLv3 and PEL
+ * @copyright  Copyright (c) elements.at New Media Solutions GmbH (https://www.elements.at)
+ * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
 namespace Elements\Bundle\ProcessManagerBundle\Controller;
@@ -20,7 +20,9 @@ use Pimcore\Bundle\AdminBundle\Helper\QueryParams;
 use Pimcore\Bundle\AdminBundle\HttpFoundation\JsonResponse;
 use Pimcore\Controller\Traits\JsonHelperTrait;
 use Pimcore\Controller\UserAwareController;
+use Pimcore\Model\Exception\NotFoundException;
 use Symfony\Component\HttpFoundation\Request;
+
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route(path: '/admin/elementsprocessmanager/callback-settings')]
@@ -28,15 +30,12 @@ class CallbackSettingsController extends UserAwareController
 {
     use JsonHelperTrait;
 
-    /**
-     * @return JsonResponse
-     */
     #[Route(path: '/save')]
-    public function saveAction(Request $request)
+    public function saveAction(Request $request): JsonResponse
     {
         try {
-            $values = json_decode((string) $request->get('values'), true, 512, JSON_THROW_ON_ERROR);
-            $settings = json_decode((string) $request->get('settings'), true, 512, JSON_THROW_ON_ERROR);
+            $values = json_decode((string)$request->get('values'), true, 512, JSON_THROW_ON_ERROR);
+            $settings = json_decode((string)$request->get('settings'), true, 512, JSON_THROW_ON_ERROR);
             if ($request->get('id')) {
                 $setting = CallbackSetting::getById($request->get('id'));
             } else {
@@ -62,6 +61,11 @@ class CallbackSettingsController extends UserAwareController
     {
         try {
             $setting = CallbackSetting::getById($request->get('id'));
+            if (!$setting) {
+                throw new NotFoundException(
+                    sprintf('Callback setting id %d', $request->get('id'))
+                );
+            }
             $setting->delete();
 
             return $this->jsonResponse(['success' => true]);
@@ -79,7 +83,7 @@ class CallbackSettingsController extends UserAwareController
         try {
             $setting = CallbackSetting::getById($request->get('id'));
             if ($setting) {
-                $setting->setId(null)->setName('Copy - ' . $setting->getName())->save();
+                $setting->setId(-1)->setName('Copy - ' . $setting->getName())->save();
 
                 return $this->jsonResponse(['success' => true]);
             } else {
@@ -116,7 +120,7 @@ class CallbackSettingsController extends UserAwareController
         $data = [];
         foreach ($list->load() as $item) {
             $tmp = $item->getObjectVars();
-            $tmp['extJsSettings'] = json_decode((string) $tmp['settings'], true, 512, JSON_THROW_ON_ERROR);
+            $tmp['extJsSettings'] = json_decode((string)$tmp['settings'], true, 512, JSON_THROW_ON_ERROR);
             $data[] = $tmp;
         }
